+++ CHANGELOG.md
@@ -9,10 +9,11 @@ Starting after Rigbox 2.2.0, this file contains a curated, chronologically order
 - tests for SignalsExp class
 - alyx warnings not thrown when database url not defined
 - new SignalsExp test GUI has numerous bugfixes and allows you to view exp panel
-- exp.trialConditions allows trial reset, similar to ConditionServer
+- exp.trialConditions allows trial reset, similar to ConditionServer `62eb9ac` 2019-11-14
 - fix for catastrophic crash when stimulus layers contain empty values
 - time signal now always updates before expStart event `cab5a2f` 2019-11-01
-- ability to hide event updates in the ExpPanel
+- ability to hide event updates in the ExpPanel `fef6ac2` 2019-11-14
+- 
 
 
 ## [2.4.0](https://github.com/cortex-lab/Rigbox/releases/tag/v2.4.0)